chore: rename createSearchSlice -> createSessionSearchSlice
