+++ packages/redux-saga/src/index.ts
@@ -1,5 +1,5 @@
 export * from "./slice/createCrudSlice";
-export * from "./slice/createSearchSlice";
+export * from "./slice/createSessionSearchSlice";
 export * from "./slice/createOptionsSlice";
 export * from "./types";
 export * from "./action/actionTypes";
